Move FontLayer to Outputs folder
