+++ PyRoute/SubsectorMap2.py
@@ -11,7 +11,7 @@ from Position.Hex import Hex
 from PyRoute.Map import GraphicMap
 from PyRoute.Galaxy import Galaxy
 from PyRoute.AllyGen import AllyGen
-from PyRoute.FontLayer import FontLayer
+from PyRoute.Outputs.FontLayer import FontLayer
 from PIL import Image, ImageDraw, ImageColor, ImageFont
